refactor(domain): move in domain the event reaction adt and add time utils
